@@ -16,11 +16,11 @@
     ],
     "config": {
         "platform": {
-            "php": "7.3.31"
+            "php": "8.2.29"
         }
     },
     "require": {
-        "php": ">=7.3.0"
+        "php": ">=8.2.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~9.0"
