@@ -9,10 +9,10 @@ use randomhost\Icinga\Check\Check;
 use randomhost\Icinga\Plugin;
 
 /**
- * Testing stub for randomhost\Icinga\Check\Base.
+ * Testing stub for {@see Base}.
  *
  * @author    Ch'Ih-Yu <chi-yu@web.de>
- * @copyright 2022 Random-Host.tv
+ * @copyright 2025 Random-Host.tv
  * @license   https://opensource.org/licenses/BSD-3-Clause BSD License (3 Clause)
  *
  * @see https://github.random-host.tv
@@ -21,17 +21,13 @@ class Stub extends Base implements Check
 {
     /**
      * Expected exit message.
-     *
-     * @var string
      */
-    private $stubMessage = '';
+    private string $stubMessage = '';
 
     /**
      * Expected exit code.
-     *
-     * @var int
      */
-    private $stubCode = Plugin::STATE_UNKNOWN;
+    private int $stubCode = Plugin::STATE_UNKNOWN;
 
     /**
      * Stub constructor.
@@ -73,8 +69,6 @@ class Stub extends Base implements Check
 
     /**
      * Executes the main Icinga check plugin logic.
-     *
-     * @return $this
      */
     protected function check(): Plugin
     {
