@@ -10,11 +10,9 @@ use randomhost\Icinga\Plugin;
  * Interface definition for Icinga check plugins.
  *
  * @author    Ch'Ih-Yu <chi-yu@web.de>
- * @copyright 2022 Random-Host.tv
+ * @copyright 2025 Random-Host.tv
  * @license   https://opensource.org/licenses/BSD-3-Clause BSD License (3 Clause)
  *
  * @see https://github.random-host.tv
  */
-interface Check extends Plugin
-{
-}
+interface Check extends Plugin {}
